mdadm.h: Add mdstat.h
Move mdstat_ent handling to separate header and include
this header directly.

No functional changes are intended.
