+++ Detail.c
@@ -23,6 +23,7 @@
  */
 
 #include	"mdadm.h"
+#include	"mdstat.h"
 #include	"xmalloc.h"
 
 #include	<ctype.h>
