+++ Grow.c
@@ -22,6 +22,7 @@
  *    Email: <neilb@suse.de>
  */
 #include	"mdadm.h"
+#include	"mdstat.h"
 #include	"dlink.h"
 #include	"xmalloc.h"
