+++ mdmon.c
@@ -64,6 +64,7 @@
 #endif
 
 #include	"mdadm.h"
+#include	"mdstat.h"
 #include	"mdmon.h"
 #include	"xmalloc.h"
